Basic-Weather-App/requirements.txt
requests

Pillow  # For image handling in Tkinter
